moved logic of geneExpression to charts and added chartTypes for geneExpression and metaboliteIntensity
