+++ client/plots/hierCluster.config.js
@@ -1,7 +1,6 @@
 import { copyMerge } from '../rx'
 import { getPlotConfig as getMatrixPlotConfig } from './matrix.config'
 import { fillTermWrapper, get$id } from '#termsetting'
-import { showGenesetEdit } from '../dom/genesetEdit.ts' // cannot use '#dom/', breaks
 import { NumericModes, TermTypes } from '../shared/terms.js'
 
 export async function getPlotConfig(opts = {}, app) {
@@ -83,132 +82,4 @@ export async function getPlotConfig(opts = {}, app) {
 
 	config.settings.matrix.maxSample = 100000
 	return config
-}
-
-export function makeChartBtnMenu(holder, chartsInstance) {
-	/*
-	holder: the holder in the tooltip
-	chartsInstance: MassCharts instance
-        termdbConfig is accessible at chartsInstance.state.termdbConfig{}
-        mass option is accessible at chartsInstance.app.opts{}
-	*/
-
-	// to fill in menu, create options in holder
-	/*
-	holder.append('div')
-		.attr('class','sja_menuoption sja_sharp_border')
-		.text('Single gene expression')
-		.on('click',()=>{
-			chartsInstance.dom.tip.hide()
-			
-		})
-		*/
-	const geneList = []
-	const app = chartsInstance.app
-	const tip = app.tip
-
-	holder.selectAll('*').remove()
-	const div = holder.append('div').style('padding', '5px')
-	const label = div.append('label')
-	label.append('span').text('Create ')
-	let name
-	const nameInput = label
-		.append('input')
-		.style('margin', '2px 5px')
-		.style('width', '210px')
-		.attr('placeholder', 'Group Name')
-		.on('input', () => {
-			name = nameInput.property('value')
-		})
-	const selectedGroup = {
-		index: 0,
-		name,
-		label: name,
-		lst: [],
-		status: 'new'
-	}
-
-	showGenesetEdit({
-		holder: holder.append('div'),
-		/* running hier clustering and the editing group is the group used for clustering
-	pass this mode value to inform ui to support the optional button "top variably exp gene"
-	this is hardcoded for the purpose of gene expression and should be improved
-	*/
-		genome: app.opts.genome,
-		geneList,
-		mode: 'expression',
-		vocabApi: app.vocabApi,
-		callback: async ({ geneList, groupName }) => {
-			if (!selectedGroup) throw `missing selectedGroup`
-			tip.hide()
-			const group = { name: groupName || name, lst: [], type: 'hierCluster' }
-			const lst = group.lst.filter(tw => tw.term.type != 'geneVariant')
-			const tws = await Promise.all(
-				geneList.map(async d => {
-					const term = {
-						gene: d.symbol || d.gene,
-						name: d.symbol || d.gene,
-						type: 'geneVariant'
-					}
-					//if it was present use the previous term, genomic range terms require chr, start and stop fields, found in the original term
-					let tw = group.lst.find(tw => tw.term.name == d.symbol || tw.term.name == d.gene)
-					if (!tw) {
-						tw = { term, q: {} }
-						tw.$id = await get$id()
-					} else if (!tw.$id) {
-						tw.$id = await get$id()
-					}
-					return tw
-				})
-			)
-
-			if (tws.length == 1) {
-				const gene = tws[0].term.gene
-				const tw = { term: { name: gene, gene, type: TermTypes.GENE_EXPRESSION }, q: { mode: NumericModes.continuous } }
-
-				app.dispatch({
-					type: 'plot_create',
-					config: {
-						chartType: 'summary',
-						term: tw
-					}
-				})
-				return
-			}
-
-			if (tws.length == 2) {
-				const gene = tws[0].term.gene
-				const tw = { term: { name: gene, gene, type: TermTypes.GENE_EXPRESSION }, q: { mode: NumericModes.continuous } }
-				const gene2 = tws[1].term.gene
-				const tw2 = {
-					term: { name: gene2, gene: gene2, type: TermTypes.GENE_EXPRESSION },
-					q: { mode: NumericModes.continuous }
-				}
-
-				app.dispatch({
-					type: 'plot_create',
-					config: {
-						chartType: 'summary',
-						term: tw,
-						term2: tw2
-					}
-				})
-				return
-			}
-			group.lst = [...lst, ...tws]
-			if (!group.lst.length) tg.splice(selectedGroup.index, 1)
-
-			// close geneset edit ui after clicking submit
-			holder.selectAll('*').remove()
-
-			app.dispatch({
-				type: 'plot_create',
-				config: {
-					chartType: 'hierCluster',
-					termgroups: [group],
-					dataType: TermTypes.GENE_EXPRESSION
-				}
-			})
-		}
-	})
 }
